+++ examples/local-ipc/src/providers.rs
@@ -6,13 +6,13 @@ pub const EXAMPLE_LOCAL_IPC_NAME: &str = "service-daemon-rs-local-ipc-example";
 pub const EXAMPLE_LOCAL_IPC_ENV: &str = "SERVICE_DAEMON_RS_LOCAL_IPC_EXAMPLE_NAME";
 
 #[provider(
-    LocalIpcListen("service-daemon-rs-local-ipc-example"),
-    env = "SERVICE_DAEMON_RS_LOCAL_IPC_EXAMPLE_NAME"
+    LocalIpcListen(EXAMPLE_LOCAL_IPC_NAME),
+    env = EXAMPLE_LOCAL_IPC_ENV
 )]
 pub struct ExampleLocalIpcListener;
 
 #[provider(
-    LocalIpcConnect("service-daemon-rs-local-ipc-example"),
-    env = "SERVICE_DAEMON_RS_LOCAL_IPC_EXAMPLE_NAME"
+    LocalIpcConnect(EXAMPLE_LOCAL_IPC_NAME),
+    env = EXAMPLE_LOCAL_IPC_ENV
 )]
 pub struct ExampleLocalIpcConnector;