+++ examples/unix-domain-socket/src/providers.rs
@@ -3,8 +3,8 @@ use service_daemon::provider;
 pub const EXAMPLE_UNIX_DOMAIN_SOCKET_PATH: &str =
     "/tmp/service-daemon-unix-domain-socket-example.socket";
 
-#[provider(UnixListen("/tmp/service-daemon-unix-domain-socket-example.socket"))]
+#[provider(UnixListen(EXAMPLE_UNIX_DOMAIN_SOCKET_PATH))]
 pub struct ExampleUnixDomainSocketListener;
 
-#[provider(UnixConnect("/tmp/service-daemon-unix-domain-socket-example.socket"))]
+#[provider(UnixConnect(EXAMPLE_UNIX_DOMAIN_SOCKET_PATH))]
 pub struct ExampleUnixDomainSocketConnector;